Adaptando para trabalhar com fundos de investimento

+++ src/main/groovy/br/com/carteira/entity/Operacao.groovy
@@ -8,7 +8,7 @@ class Operacao {
     TipoOperacaoEnum tipoOperacao
     LocalDate data
     Ativo ativo
-    Integer qtde
+    BigDecimal qtde
     BigDecimal valorTotalOperacao
     BigDecimal custoMedioVenda
     BigDecimal resultadoVenda
